requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
python-dotenv
requests
werkzeug
gunicorn
pandas